+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.3-DEV] - 2020-09-17
 - do not add block templates to TemplateLoader
+- fixed bundle order for template loading
 
 ## [0.1.2] - 2020-09-17
 - fixed wrong method call lead to navigation templates not rendered