+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.3-DEV] - 2020-09-17
-- switch to symfony PSR-6 cache due strange errors in contao 4.4 with PSR-16 filesystem cache.
+## [0.1.3] - 2020-09-17
+- switch to symfony PSR-6 cache due strange errors in contao 4.4 with PSR-16 filesystem cache
+- added ContaoWidgetNormalizer
+- made normalization contao 4.4 compatible
 - twig templates now only added in backend to TemplateLoader
 - fixed bundle order for template loading
+- fixed options not correctly loaded from widget
 
 ## [0.1.2] - 2020-09-17
 - fixed wrong method call lead to navigation templates not rendered